refactor(storage): split `MD5Hash()` internal function
Use separate files(s) for this function. That would make it easier to
provide a different implementation.
