+++ google/cloud/storage/google_cloud_cpp_storage.bzl
@@ -83,6 +83,7 @@ google_cloud_cpp_storage_hdrs = [
     "internal/lifecycle_rule_parser.h",
     "internal/logging_stub.h",
     "internal/make_jwt_assertion.h",
+    "internal/md5hash.h",
     "internal/metadata_parser.h",
     "internal/notification_metadata_parser.h",
     "internal/notification_requests.h",
@@ -200,6 +201,7 @@ google_cloud_cpp_storage_srcs = [
     "internal/lifecycle_rule_parser.cc",
     "internal/logging_stub.cc",
     "internal/make_jwt_assertion.cc",
+    "internal/md5hash.cc",
     "internal/metadata_parser.cc",
     "internal/notification_metadata_parser.cc",
     "internal/notification_requests.cc",
