+++ google/cloud/storage/google_cloud_cpp_storage.cmake
@@ -132,6 +132,8 @@ add_library(
     internal/logging_stub.h
     internal/make_jwt_assertion.cc
     internal/make_jwt_assertion.h
+    internal/md5hash.cc
+    internal/md5hash.h
     internal/metadata_parser.cc
     internal/metadata_parser.h
     internal/notification_metadata_parser.cc
@@ -482,6 +484,7 @@ if (BUILD_TESTING)
         internal/impersonate_service_account_credentials_test.cc
         internal/logging_stub_test.cc
         internal/make_jwt_assertion_test.cc
+        internal/md5hash_test.cc
         internal/metadata_parser_test.cc
         internal/notification_requests_test.cc
         internal/object_acl_requests_test.cc
