+++ google/cloud/storage/internal/openssl_util.cc
@@ -15,7 +15,6 @@
 #include "google/cloud/storage/internal/openssl_util.h"
 #include "google/cloud/internal/base64_transforms.h"
 #include "google/cloud/internal/openssl_util.h"
-#include <openssl/evp.h>
 #include <memory>
 
 namespace google {
@@ -54,16 +53,6 @@ StatusOr<std::vector<std::uint8_t>> UrlsafeBase64Decode(
     b64str.append("=");
   }
   return Base64Decode(b64str);
-}
-
-std::vector<std::uint8_t> MD5Hash(absl::string_view payload) {
-  std::array<unsigned char, EVP_MAX_MD_SIZE> digest;
-
-  unsigned int size = 0;
-  EVP_Digest(payload.data(), payload.size(), digest.data(), &size, EVP_md5(),
-             nullptr);
-  return std::vector<std::uint8_t>{digest.begin(),
-                                   std::next(digest.begin(), size)};
 }
 
 }  // namespace internal
