+++ google/cloud/storage/internal/openssl_util_test.cc
@@ -80,37 +80,6 @@ TEST(OpensslUtilTest, Base64DecodePadding) {
               ElementsAre('A', 'B', 'C', 'D'));
 }
 
-TEST(OpensslUtilTest, MD5HashEmpty) {
-  // I used this command to get the expected value:
-  // /bin/echo -n "" | openssl md5
-  auto const expected =
-      std::vector<std::uint8_t>{0xd4, 0x1d, 0x8c, 0xd9, 0x8f, 0x00, 0xb2, 0x04,
-                                0xe9, 0x80, 0x09, 0x98, 0xec, 0xf8, 0x42, 0x7e};
-
-  // There are many ways to represent the "empty" string:
-  std::vector<std::vector<std::uint8_t>> const values{
-      MD5Hash({}),
-      MD5Hash(nullptr),
-      MD5Hash(""),
-      MD5Hash(absl::string_view{}),
-  };
-
-  for (auto const& actual : values) {
-    EXPECT_THAT(actual, ElementsAreArray(expected));
-  }
-}
-
-TEST(OpensslUtilTest, MD5HashSimple) {
-  auto const actual = MD5Hash("The quick brown fox jumps over the lazy dog");
-  // I used this command to get the expected value:
-  // /bin/echo -n "The quick brown fox jumps over the lazy dog" |
-  //     openssl md5 -binary | openssl base64
-  auto const expected =
-      std::vector<std::uint8_t>{0x9e, 0x10, 0x7d, 0x9d, 0x37, 0x2b, 0xb6, 0x82,
-                                0x6b, 0xd8, 0x1d, 0x35, 0x42, 0xa4, 0x19, 0xd6};
-  EXPECT_THAT(actual, ElementsAreArray(expected));
-}
-
 }  // namespace
 }  // namespace internal
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
